+++ yt_queue/filters.py
@@ -1,3 +1,4 @@
+import math
 import re
 
 def filter_videos(data, filters):
@@ -7,6 +8,10 @@ def filter_videos(data, filters):
         videos = _filter_by_status(videos, filters['status'])
     if 'title' in filters:
         videos = _filter_by_title(videos, filters['title'])
+    if 'min-duration' in filters:
+        videos = _filter_by_duration(videos, filters['min-duration'], math.inf)
+    if 'max-duration' in filters:
+        videos = _filter_by_duration(videos, 0, filters['max-duration'])
     if 'custom' in filters:
         videos = _filter_by_custom(videos, filters['custom'])
 
@@ -25,6 +30,12 @@ def _filter_by_title(videos, title_pattern):
     for video in videos:
         if 'title' in video and title_re.match(video['title']):
             yield video
+
+def _filter_by_duration(videos, min_duration, max_duration):
+    for video in videos:
+        if 'duration' in video:
+            if video['duration'] >= min_duration and video['duration'] <= max_duration:
+                yield video
 
 def _filter_by_custom(videos, fn):
     for video in videos: